Move current algorithm to separate file.
